requirements.txt
fastapi==0.63.0
python-telegram-bot==20.8
requests==2.31.0
pytz==2022.1